+++ apps/ced-card-request-fe/src/features/confirmation/reducer.ts
@@ -1,0 +1,36 @@
+import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
+import { ConfermaDomandaRequest } from '../../core/api/generated/model';
+
+const initialState: ConfermaDomandaRequest = {} as ConfermaDomandaRequest;
+
+const confirmRequestFormSlice = createSlice({
+  name: 'confirmRequestForm',
+  initialState,
+  reducers: {
+    setField: (
+      state,
+      action: PayloadAction<{
+        field: keyof ConfermaDomandaRequest;
+        value: ConfermaDomandaRequest[keyof ConfermaDomandaRequest];
+      }>,
+    ) => {
+      const { field, value } = action.payload;
+      return {
+        ...state,
+        [field]: value,
+      };
+    },
+    setForm: (
+      state,
+      action: PayloadAction<Partial<ConfermaDomandaRequest>>,
+    ) => ({
+      ...state,
+      ...action.payload,
+    }),
+    resetForm: () => initialState,
+  },
+});
+
+export const { setField, setForm, resetForm } = confirmRequestFormSlice.actions;
+
+export const confirmRequestFormReducer = confirmRequestFormSlice.reducer;